+++ note-app/archivedNotesPage/archivedNotesPage.html
@@ -4,14 +4,14 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./styles/tables.css">
+    <link rel="stylesheet" href="../styles/tables.css">
     <title>Document</title>
 </head>
 <body>
     <section>
-        <a href="./index.html"><button>Go back</button></a>
+        <a href="../index.html"><button>Go back</button></a>
         <table class="table">
-            <tbody class="table-body" id="main-board">
+            <tbody id="archived-notes-board">
                 <tr class="row-main">
                     <th>Name</th>
                     <th>Created</th>
@@ -25,6 +25,6 @@
             </tbody>
         </table>
     </section>
-    <script src="./scripts/archivedNotes.js"></script>
+    <script type="module" src="./archivedNotes.js"></script>
 </body>
 </html>
