+++ note-app/createNotePage/createNote.js
@@ -1,9 +1,5 @@
+import { getMentionedDates } from "../scripts/modules/index.js";
 const addNoteForm = document.querySelector('#add-note-form')
-
-const getMentionedDates = (text) => {
-    const dayMonthYearRegex = /^(0?[1-9]|[12][0-9]|3[01])[\/\-](0?[1-9]|1[012])[\/\-]\d{4}$/
-    return text.replace(/[^a-zA-Z0-9-/\ ]/g, "").split(" ").filter(item=>item.match(dayMonthYearRegex))
-}
 
 addNoteForm.addEventListener('submit',(event)=>{
     const noteId = "id" + Math.random().toString(16).slice(2)
