+++ note-app/editNotePage/editNote.js
@@ -1,3 +1,5 @@
+import { getMentionedDates } from "../scripts/modules/index.js";
+
 const nameInput = document.querySelector('#name')
 const editNoteForm = document.querySelector('#edit-note-form')
 const categorySelection = document.querySelector('#category')
@@ -14,11 +16,6 @@ function getNoteData() {
         textContentInput.value = noteData.text_content
 }
 getNoteData()
-
-const getMentionedDates = (text) => {
-    const dayMonthYearRegex = /^(0?[1-9]|[12][0-9]|3[01])[\/\-](0?[1-9]|1[012])[\/\-]\d{4}$/
-    return text.replace(/[^a-zA-Z0-9-/\ ]/g, "").split(" ").filter(item=>item.match(dayMonthYearRegex))
-}
 
 editNoteForm.addEventListener('submit',(event)=>{
     const urlSearchParams = new URLSearchParams(window.location.search);
