+++ note-app/index.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <article>
-        <a href="./archivedNotes.html"><button>Archived notes</button></a>
+        <a href="./archivedNotesPage/archivedNotesPage.html"><button>Archived notes</button></a>
         <section>
             <table>
                 <tbody id="main-board">
@@ -26,7 +26,7 @@
                     </tr>
                 </tbody>
             </table>
-            <a href="./inputForm.html"><button>Add note</button></a>
+            <a href="./createNotePage/createNotePage.html"><button>Add note</button></a>
         </section>
         <section>
             <h2>Stats</h2>
@@ -39,7 +39,7 @@
             </table>
         </section>
     </article>
-    <script src="./scripts/index.js"></script>
+    <script type="module" src="./scripts/index.js"></script>
     <script src="./scripts/noteStats.js"></script>
 </body>
 </html>
